Move extensions into nested namespace and use global usings to have them available everywhere
