+++ Mutagen.Bethesda.Analyzers.Skyrim/Extensions/LocationExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using Mutagen.Bethesda.Plugins;
 using Mutagen.Bethesda.Plugins.Cache;
 using Mutagen.Bethesda.Skyrim;
-namespace Mutagen.Bethesda.Analyzers.Skyrim;
+namespace Mutagen.Bethesda.Analyzers.Skyrim.Extensions;
 
 public static class LocationExtensions
 {
