+++ Mutagen.Bethesda.Analyzers.Skyrim/Extensions/WorldspaceExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Mutagen.Bethesda.Plugins.Cache;
 using Mutagen.Bethesda.Skyrim;
-namespace Mutagen.Bethesda.Analyzers.Skyrim;
+namespace Mutagen.Bethesda.Analyzers.Skyrim.Extensions;
 
 public static class WorldspaceExtensions
 {
