+++ Mutagen.Bethesda.Analyzers.Skyrim/GlobalUsings.cs
@@ -1,0 +1,1 @@
+﻿global using Mutagen.Bethesda.Analyzers.Skyrim.Extensions;
